requirements.txt
streamlit==1.33.0
streamlit-audiorec==0.1.3
faster-whisper==0.10.0     # requires av 11.*
torch==2.2.1
pydub==0.25.1
python-Levenshtein==0.24.0
sentence-transformers==2.6.1
gTTS==2.5.1
av==11.0.0                  # wheel exists for cp311 arm64
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.2           # resampling helper
numba==0.59.1             # librosa runtime dependency
requests==2.32.3
azure-cognitiveservices-speech==1.37.0
webrtcvad==2.0.10